+++ src/game/tests.rs
@@ -467,6 +467,7 @@ mod vintage_cube_once_upon_a_time;
 mod vintage_cube_oracle;
 mod vintage_cube_orcish_bowmasters;
 mod vintage_cube_otawara;
+mod vintage_cube_otharri;
 mod vintage_cube_ouroboroid;
 mod vintage_cube_overlord;
 mod vintage_cube_paradoxical_outcome;